Reorganize run utils
