+++ lib/run/browsers/index.js
@@ -1,8 +1,8 @@
 import logger from '@util/logger';
 
+import resolveLocal from '../util/resolve-local';
+import { getDefaultBrowser } from '../util/browser';
 import BaseBrowser from './base';
-import { maybeResolveLocal } from '../util/common';
-import { getDefaultBrowser } from '../util/browsers';
 
 const { assign } = Object;
 
@@ -17,10 +17,7 @@ const { assign } = Object;
  * export is not an instance of the base browser class
  */
 export function requireBrowser(name) {
-  let module = maybeResolveLocal('browsers', name);
-
-  if (!module) throw new Error(`cannot find browser "${name}"`);
-
+  let module = resolveLocal('browser', name);
   let Browser = require(module).default;
 
   if (!(Browser && Browser.prototype instanceof BaseBrowser)) {
