+++ lib/run/plugins/adapter.js
@@ -1,15 +1,7 @@
 import BasePlugin from './base';
-import { maybeResolveLocal } from '../util/common';
+import resolveLocal from '../util/resolve-local';
 
 const { assign } = Object;
-
-export function resolveAdapterPath(name) {
-  let module = maybeResolveLocal('adapters', name);
-
-  if (!module) throw new Error(`cannot find adapter "${name}"`);
-
-  return module;
-}
 
 export default class AdapterPlugin extends BasePlugin {
   static name = 'adapter';
@@ -27,7 +19,7 @@ export default class AdapterPlugin extends BasePlugin {
 
     if (!options.module) {
       options = assign({}, options, {
-        module: resolveAdapterPath(options.name)
+        module: resolveLocal('adapter', options.name)
       });
     }
