+++ lib/run/plugins/index.js
@@ -1,13 +1,10 @@
+import resolveLocal from '../util/resolve-local';
 import BasePlugin from './base';
-import { maybeResolveLocal } from '../util/common';
 
 const { assign } = Object;
 
 export function requirePlugin(name) {
-  let module = maybeResolveLocal('plugins', name);
-
-  if (!module) throw new Error(`cannot find plugin "${name}"`);
-
+  let module = resolveLocal('plugin', name);
   let Plugin = require(module).default;
 
   if (!(Plugin && Plugin.prototype instanceof BasePlugin)) {
