+++ lib/run/plugins/serve.js
@@ -1,6 +1,6 @@
 import BasePlugin from './base';
 import ChildProcess from '../process';
-import { request } from '../util/network';
+import request from '../util/request';
 
 const { assign } = Object;
