+++ lib/run/process.js
@@ -1,11 +1,11 @@
-import { spawn, spawnSync } from 'child_process';
+import { spawn } from 'child_process';
 import { existsSync } from 'fs-extra';
 import { when } from '@bigtest/convergence';
 
-import { hasDescriptor } from './util/common';
+import { hasDescriptor } from './util/descriptors';
+import binExists from './util/bin-exists';
 
 const { assign } = Object;
-const binExists = bin => !spawnSync('which', [bin]).status;
 
 /**
  * Wraps a child process so that it can be started with async start
