+++ lib/run/reporters/index.js
@@ -1,5 +1,5 @@
+import resolveLocal from '../util/resolve-local';
 import BaseReporter from './base';
-import { maybeResolveLocal } from '../util/common';
 
 const { assign } = Object;
 
@@ -14,10 +14,7 @@ const { assign } = Object;
  * export is not an instance of the base reporter class
  */
 export function requireReporter(name) {
-  let module = maybeResolveLocal('reporters', name);
-
-  if (!module) throw new Error(`cannot find reporter "${name}"`);
-
+  let module = resolveLocal('reporter', name);
   let Reporter = require(module).default;
 
   if (!(Reporter && Reporter.prototype instanceof BaseReporter)) {
