+++ lib/run/servers/sockets.js
@@ -1,7 +1,7 @@
 import url from 'url';
 import EventEmitter from 'events';
 import WebSocket from 'ws';
-import { parseBrowserName } from '../util/browsers';
+import { parseBrowserName } from '../util/browser';
 
 function uid() {
   return (uid.uid = (uid.uid || 0) + 1).toString().padStart(3, '0');
