+++ lib/run/state/store.js
@@ -1,4 +1,4 @@
-import { getDescriptors } from '../util/common';
+import { getDescriptors } from '../util/descriptors';
 
 const {
   assign,
