+++ lib/run/util/descriptors.js
@@ -1,6 +1,3 @@
-import path from 'path';
-import { existsSync } from 'fs-extra';
-
 const {
   assign,
   getOwnPropertyDescriptor,
@@ -45,24 +42,4 @@ export function getDescriptors(obj) {
 
   delete descr.constructor;
   return descr;
-}
-
-/**
- * Checks for the a file in a local directory within this package
- * (browsers, adapters, reporters). If it doesn't exist locally,
- * attempt to resolve the absolute path using `require.resolve`.
- *
- * @private
- * @param {String} dir - Local directory to look in
- * @param {String} name - Name of local file, or module path
- * @returns {String} the module path, or null if not found
- */
-export function maybeResolveLocal(dir, name) {
-  let local = path.join(__dirname, `../${dir}/${name}.js`);
-  let module = existsSync(local) ? local : null;
-
-  // if not local, try to resolve the name directly
-  try { module = module || require.resolve(name); } catch (e) {}
-
-  return module;
 }
